\name{RcmdrPlugin.MA-package}
\alias{RcmdrPlugin.MA-package}
\docType{package}
\title{Graphical User Interface for Meta-Analysis in R
}
\description{ This package provides an Rcmdr plug-in for conducting a meta-analysis.  This package will run many of the common (and not so common, e.g. several procedures for handling ES dependencies) meta-analytic functions available in R though a familiar Graphical User Interface (GUI) environment. 
\tabular{ll}{
Package: \tab RcmdrPlugin.MA-package\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2014-09-21\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{AC Del Re 

Maintainer: AC Del Re <acdelre@gmail.com>
}
\references{ AC Del Re \& William T. Hoyt (2010). \emph{MAd: Meta-Analysis with Mean Differences}. R package version 0.9. http://CRAN.R-project.org/package=MAd

AC Del Re \& William T. Hoyt (2010). \emph{MAc: Meta-Analysis with Correlations}. R package version 1.1. http://CRAN.R-project.org/package=MAc

Wolfgang Viechtbauer (2010). Conducting meta-analyses in R with the metafor package. Journal of Statistical Software, 36(3), 1-48. URL http://www.jstatsoft.org/v36/i03/.
 
}
\examples{
## Installing the RcmdrPlugin.MA package (at command prompt):  
#    install.packages('RcmdrPlugin.MA')               

## Once this is installed, it can be loaded by running the command:   
#    library(RcmdrPlugin.MA)


}  
\keyword{ package }
\seealso{
\bold{MAd} package:  \url{http://CRAN.R-project.org/package=MAd};
\bold{MAc} package:  \url{http://CRAN.R-project.org/package=MAc};
\bold{metafor} package:  \url{http://www.jstatsoft.org/v36/i03/}
}